refactor: 상품 리포지터리를 분리하라
