+++ src/main/java/kitchenpos/products/infra/JpaProductRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.products.domain;
+package kitchenpos.products.infra;
 
+import kitchenpos.products.domain.Product;
+import kitchenpos.products.domain.tobe.ProductRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
