+++ src/test/java/kitchenpos/menus/application/MenuServiceTest.java
@@ -7,8 +7,8 @@ import kitchenpos.menus.domain.MenuRepository;
 import kitchenpos.products.application.FakeBanWordFilter;
 import kitchenpos.products.application.InMemoryProductRepository;
 import kitchenpos.products.domain.Product;
-import kitchenpos.products.domain.ProductRepository;
 import kitchenpos.products.domain.tobe.BanWordFilter;
+import kitchenpos.products.domain.tobe.ProductRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
