+++ src/test/java/kitchenpos/products/application/InMemoryProductRepository.java
@@ -1,9 +1,14 @@
 package kitchenpos.products.application;
 
 import kitchenpos.products.domain.Product;
-import kitchenpos.products.domain.ProductRepository;
+import kitchenpos.products.domain.tobe.ProductRepository;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.UUID;
 import java.util.stream.Collectors;
 
 public class InMemoryProductRepository implements ProductRepository {
